% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfaser-package.R
\docType{package}
\name{pfaser-package}
\alias{pfaser-package}
\alias{pfaser}
\title{The 'pfaser' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
